An automated Canny edge detection function is defined where the function requires a single argument, i.e. image, which is a single-channel image. The sigma is used to vary the percentage thresholds that are determined based on the median.
Initially, the computing of median of the pixel intensities in the image is done.

Then, this median value is used to construct two thresholds, lower and upper. These thresholds are constructed based on the +/- percentages controlled by the sigma argument.

A lower value of sigma indicates a tighter threshold, whereas a larger value of sigma gives a wider threshold. In general, the sigma value will not have to be changed often. Simply, a single, default sigma value needs to be selected and applied to the entire dataset of images.