In this code, the color and texture features of an image are normalized (on the same scale) in order to merge both of them for creating a multi-feature image. This multi-feature image is being used as an input to the kmeans model to classify each pixel into different class. The road pixels are then highlighted with randomly generated colors.
Color feature normalization: The image is converted to HSV, split into h,s,v for normalizing each channel seperately, and finally stacked together again, for displaying a normalized HSV image.

Texture feature normalization: The build_filters() function crates a set of Gabor filters to extract texture features from an image. It does this by creating a filter for each angle (in radians) in a range of 0 to pi, with a step of pi/16, using the 'cv2.getGaborKernel' function. The size of the filter os set to 31x31 and the other parameters for the Gabor filter (sigma, lambda, gamma) are also set. The filter is then normalized by dividing it with 1.5 times the sum of its elements. All these filters are then appended to a list and returned.
 The process() function takes an image and the set of filters and performs feature extraction. It applies each filter to the image and accumulates the results in a matrix. The 'np.maximum' function is then used to update the accumulated matrix with the maximum values of the filtered image and the accumulated matrix, to get the final result. This final result is returned as the texture features.

Finally, the extract_road() function combines the color and texture features, applies K-means clustering on the combined features, and returns an image with the road pixels highlighted.