Implementation of Perspective transform
Implemented the algorithm for applying perspective transform. Initially a trapezoid ROI is defined, which acts as the pilot's view from the airplane. This trapezoid-like perspective is warped into a rectangle-like perspective. The texture feature normalization is consequently applied on the warped image for testing purpose. Once the desired output can be attained from the warped image, it can be unwarped into its original structure, which is also implemented in this program.